Fix jMQTT::MQTTCLIENT* in HTML Fix daemon up/down logs Move jMqttClient to a separte .py file Code factor in jMQTT::on_daemon_connect()
